backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic>=2.0
aiofiles
sqlmodel
pymysql
cryptography
bcrypt
python-dotenv
